+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ### Latest/Nightly (this branch)
+- Feature
+    - User Messages Tool now lists tick and time.
 - Internal
     - set source-demo-tool version to 0.5.0